Refactor models governing Elasticsearch
Previously, knowledge about how to structure an Elasticsearch query
was scattered among several related models (SearchesModels,
TermSearch, TermFilter, etc.) This made it challenging to restructure
the search definition, which will be required when we upgrade to ES 6.x
and then 7.x. It was also unclear what the responsibilities of the
functions within this universe were, due to lack of commenting and
holdovers from the original ES 2.x/Tire configuration.

This refactor:
* Consolidates knowledge of the Elasticsearch DSL in SearchesModels. It
  is now the only model which knows the entire shape of the query. Other
  models only supply values for particular keys in the query hash, and
  need only to know the structure of their value, not where they belong
  in the overall data structure.
* Removes unused functionality. (e.g. we carefully constructed
  search_config[:filters], and then did not assign it to anything inside
  the search_definition actually sent to Elasticsearch!)
* Creates a seam between #prepare and #search in SearchesModels. This
  lets us test the search definition constructed by #prepare without
  needing to fire off a search (aka we don't have to run or mock out
  Elasticsearch in the model test). This will facilitate the upgrade,
  since we should just need to update the DSL syntax in
  spec/models/elasticsearch/searches_models_spec.rb, and then ensure
  that those tests pass.
* Organizes all the elasticsearch-related models into a subdirectory of
  app/models (and spec/models), for readability, and autoloads them, for
  convenience.

Unfortunately, because this changes the responsibilities of several
classes, it is not an entirely clean refactor; I had to change the tests
as I went. I have tried to preserve the coverage and intent. I did not
change any of the integration tests, so insofar as we tested user-facing
behavior, it has not changed.

+++ spec/support/elasticsearch.rb
@@ -57,10 +57,12 @@ RSpec.configure do |config|
     config.before :each, type: type do
       searchable_models.each do |model|
         begin
-          model.__elasticsearch__.client.delete_by_query(
-            index: model.__elasticsearch__.index_name
-          )
-          model.__elasticsearch__.delete_index!
+          if model.__elasticsearch__.index_exists? index: model.__elasticsearch__.index_name
+            model.__elasticsearch__.client.delete_by_query(
+              index: model.__elasticsearch__.index_name, q: '.'.freeze
+            )
+            model.__elasticsearch__.delete_index!
+          end
         rescue Elasticsearch::Transport::Transport::Errors::NotFound; end
 
         begin
